+++ engine/core/io/IO.h
@@ -62,7 +62,7 @@ namespace Echo
 
 		// 获取指定后缀名的所有文件(.bmp|.tga)
 		void listFilesWithExt( StringArray& oFiles, const char* extWithDot);
-		int getArchiveCount() const { return m_resourceIndexCaseSensitive.size(); }
+		int getArchiveCount() const { return (int)m_resourceIndexCaseSensitive.size(); }
 
 		// 获取指定目录和指定后缀的资源存档
 		void listFileWithPathAndExt(StringArray& oFiles, const char* path, const char* extWithDot);
